+++ src/vs/editor/contrib/inlineCompletions/browser/ghostText.ts
@@ -7,8 +7,7 @@ import { Emitter, Event } from 'vs/base/common/event';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { IActiveCodeEditor } from 'vs/editor/browser/editorBrowser';
 import { EditorOption } from 'vs/editor/common/config/editorOptions';
-import { Position } from 'vs/editor/common/core/position';
-import { IRange, Range } from 'vs/editor/common/core/range';
+import { applyEdits } from 'vs/editor/contrib/inlineCompletions/browser/utils';
 
 export class GhostText {
 	public static equals(a: GhostText | undefined, b: GhostText | undefined): boolean {
@@ -28,6 +27,9 @@ export class GhostText {
 			this.parts.every((part, index) => part.equals(other.parts[index]));
 	}
 
+	/**
+	 * Only used for testing/debugging.
+	*/
 	render(documentText: string, debug: boolean = false): string {
 		const l = this.lineNumber;
 		return applyEdits(documentText,
@@ -62,44 +64,6 @@ export class GhostText {
 	}
 }
 
-class PositionOffsetTransformer {
-	private readonly lineStartOffsetByLineIdx: number[];
-
-	constructor(text: string) {
-		this.lineStartOffsetByLineIdx = [];
-		this.lineStartOffsetByLineIdx.push(0);
-		for (let i = 0; i < text.length; i++) {
-			if (text.charAt(i) === '\n') {
-				this.lineStartOffsetByLineIdx.push(i + 1);
-			}
-		}
-	}
-
-	getOffset(position: Position): number {
-		return this.lineStartOffsetByLineIdx[position.lineNumber - 1] + position.column - 1;
-	}
-}
-
-function applyEdits(text: string, edits: { range: IRange; text: string }[]): string {
-	const transformer = new PositionOffsetTransformer(text);
-	const offsetEdits = edits.map(e => {
-		const range = Range.lift(e.range);
-		return ({
-			startOffset: transformer.getOffset(range.getStartPosition()),
-			endOffset: transformer.getOffset(range.getEndPosition()),
-			text: e.text
-		});
-	});
-
-	offsetEdits.sort((a, b) => b.startOffset - a.startOffset);
-
-	for (const edit of offsetEdits) {
-		text = text.substring(0, edit.startOffset) + edit.text + text.substring(edit.endOffset);
-	}
-
-	return text;
-}
-
 export class GhostTextPart {
 	constructor(
 		readonly column: number,
@@ -117,7 +81,6 @@ export class GhostTextPart {
 			this.lines.every((line, index) => line === other.lines[index]);
 	}
 }
-
 
 export interface GhostTextWidgetModel {
 	readonly onDidChange: Event<void>;
